Finished code base, testing...

+++ cs-355-lab-1/src/cs355/solution/model/shapes/Square.java
@@ -16,7 +16,7 @@ public class Square extends Rectangle
 		return width;
 	}
 
-	public void setSize(int size)
+	public void setSize(float size)
 	{
 		super.setWidth(size);
 		super.setHeight(size);
